feat: Rename SizedWrapper to KnownSizeWrapper.
