+++ tui-popup/src/lib.rs
@@ -21,14 +21,14 @@
 //! # Feature flags
 #![doc = document_features::document_features!()]
 
+mod known_size;
+mod known_size_wrapper;
 mod popup;
 mod popup_state;
-mod sized_widget;
-mod sized_wrapper;
 
 pub use crate::{
+    known_size::KnownSize,
+    known_size_wrapper::KnownSizeWrapper,
     popup::Popup,
     popup_state::{DragState, PopupState},
-    sized_widget::KnownSize,
-    sized_wrapper::SizedWrapper,
 };
